Macros: move macros table to component
